refactor(cache): split cache reporting helpers (#877)
